Extend amec sensor accumulator to 64-bit. Remove unused sensor test field.
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/19835
Tested-by: FSP CI Jenkins

+++ src/occ/amec/amec_wof.c
@@ -607,7 +607,7 @@ void amec_wof_common_steps(void)
 
     // Acquire important sensor data
     l_temp = AMECSENSOR_PTR(TEMP2MSP0)->sample;
-    l_accum = AMECSENSOR_PTR(CUR250USVDD0)->accumulator;
+    l_accum = (uint32_t)AMECSENSOR_PTR(CUR250USVDD0)->accumulator;
     l_v_chip = g_amec->wof.v_chip; //from amec_update_wof_sensors()
 
     if (g_amec->wof.state != AMEC_WOF_NO_CORE_CHANGE)
